pyproject.toml
[project]
name = "rust-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "async-fastapi-jwt-auth>=0.6.6",
    "fastapi>=0.120.4",
    "passlib>=1.7.4",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.3",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]
